06_clip_interrogator_app/requirements.txt
gunicorn
flask
Pillow
requests
tqdm
open_clip_torch
blip-vit
